Improve usability, especially of the "Remove" button: For system-defined mimetypes, don't offer to "Remove", but only to "Revert" to the system mimetype. With full tooltip+whatsThis in both cases, removing and reverting.
In theory the user should be able to make a system mimetype unused for local files
by clearing its patterns list, but that is a missing feature in shared-mime-info
(no patterns -> the global ones are used again), I just emailed the xdg list about it.
Meanwhile you have to add a dummy pattern (*.willneverexist) to achieve this.
CCBUG: 184181

svn path=/trunk/KDE/kdebase/runtime/; revision=942204