Merge branch 'nepomuk/tagsKioSlave'
This introduces a new kioslave which works on the protocol 'tags'. It allows
the users to browse their files via tags, and allows them to rename tags.

It additionally allows user to copy files to a tag folder, and have those
tags applied to that file. This feature is currently disabled because the
user could loose their files if they move the file instead of copying.