[kcmshell] Use desktop file name of module as app's desktop file
Summary:
By default kcmshell5 registers as org.kde.kcmshell5, but such a desktop
file does not exist. Which means that on Wayland neither window manager
nor task manager is able to provide a useful window icon for
configuration window.

This change tries to locate the desktop file of the module passed to
kcmshell5. If it is found it is set as the application's desktop file
name. Thus the compositor and task manager actually know the desktop
file which was used to start the kcmshell5 instance.

This is only done if kcmshell5 is invoked for only one module. In case
of multiple modules the application invoking kcmshell5 should pass a
proper desktop file name through the command line argument. And if the
desktopfile got specified already it is not updated from the module.

Test Plan: KWin Effects KCM shows proper icon on Wayland

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D3161