Fix KIOClient non-interactive mode handling renamed prompts
Summary:
In the kdelibs to KF5 transition a factory class was added for loading
the rename prompts to make KIO gui-less.

In jobs "if (isInteractive)" was changed everywhere into "if
(uiDelegateExtension)".

The former was false if we called KJob::setUiDelegate(nullptr). The
latter is only false if the calling code explicitly calls the new
KIO::Job::setUiDelegateExtension(nullptr) as well.

This left kioclient5 in the very weird state of non-interactive mode
still having rename dialogs but not error messages.

Now non-interactive mode will simply error if it hits a conflict and
stop as before.

Test Plan:
Copied a conflict. Got a warning and an exit.
Interactive mode remains the same

Reviewers: #plasma, dfaure

Reviewed By: dfaure

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D17703