Fix logic error in kioclient noninteractive argument handling
Summary:
If noninteractive /is/ set s_interactive should be false.
When s_interactive is true we show dialog and kuiserver notifications.

This was a porting bug.

Given fixing would result in a behavioural change, the default behaviour
is swapped to remain the same as the previous broken version and an
explicit --interactive flag is added.

Test Plan:
Ran a test that should produce an error without passing --noninteractive
It produced an error notification.

Reviewers: #plasma, nerdopolist, dfaure

Reviewed By: dfaure

Subscribers: broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D17652