Add utility to inhibit various actions
People want to use inhibitions with scripts, which is perfectly
sensible.

Most our inhibitors exit when the calling process quits, which is
something we definitely want, but it fails when we have people making
calls with qdbus/dbus-send and wondering why their inhibitions don't
work.

This tiny script puts surfaces all inhibitors in a way that's relatively
easy to use from shell scripts.

CCBUG: 424288
CCBUG: 422436