kstart: Remove support for specifying custom properties
While specifying properties in kstart is a powerful feature, it's also
an obscure feature. Most users are probably not even aware of it (I, as
a dev, discovered it only after digging through some kwin code).

In its current state, it depends a bit on X11. Making it work on Wayland
is doable, but I don't think it's worth the effort simply because of how
much this feature is niche.