kio-open: Special-case thunderlinks
They often are shipped with 2 // making the payload of the links the
hostname which, per-spec, is case insensitive and thus QUrl leverages
this property.
This breaks thunderlinks because they encode information in the link
that is case sensitive.
If we put it at the path section, then they seem to work fine.

CCBUG: 429408